minimal viable product

+++ handlers/apptHandlers.go
@@ -2,6 +2,7 @@ package handlers
 
 import (
 	"encoding/json"
+	"fmt"
 	"net/http"
 
 	"github.com/bmw2621/fnnp/db"
@@ -32,26 +33,42 @@ func GetApptByID(w http.ResponseWriter, r *http.Request){
 func DeleteApptByID(w http.ResponseWriter, r *http.Request){
 	w.Header().Set("Content-Type", "application/json")
 	params := mux.Vars(r)
-	var appt models.Appointment
-	db.Db.Delete(&appt, params["id"])
-	json.NewEncoder(w).Encode(appt)
+	token := r.URL.Query().Get("t")
+	if ValidateToken(token){
+		var appt models.Appointment
+		db.Db.Delete(&appt, params["id"])
+		fmt.Printf("Deleting: %v\n", params["id"])
+		json.NewEncoder(w).Encode(appt)
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+	}
 }
 
 // UpdateApptByID handles API calls to update a record in the database
 func UpdateApptByID(w http.ResponseWriter, r *http.Request){
 	params := mux.Vars(r)
-	var appt models.Appointment
-	db.Db.First(&appt, params["id"])
-	json.NewDecoder(r.Body).Decode(&appt)
-	db.Db.Save(&appt)
-	json.NewEncoder(w).Encode(appt)
+	token := r.URL.Query().Get("t")
+	if ValidateToken(token){
+		var appt models.Appointment
+		db.Db.First(&appt, params["id"])
+		json.NewDecoder(r.Body).Decode(&appt)
+		db.Db.Save(&appt)
+		json.NewEncoder(w).Encode(appt)
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+	}
 }
 
 // CreateAppt handles API calls to create a new appoint record in the database
 func CreateAppt(w http.ResponseWriter, r *http.Request){
 	w.Header().Set("Content-Type", "application/json")
-	var appt models.Appointment
-	json.NewDecoder(r.Body).Decode(&appt)
-	db.Db.Create(&appt)
-	json.NewEncoder(w).Encode(appt)
+	token := r.URL.Query().Get("t")
+	if ValidateToken(token){
+		var appt models.Appointment
+		json.NewDecoder(r.Body).Decode(&appt)
+		db.Db.Create(&appt)
+		json.NewEncoder(w).Encode(appt)
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+	}
 }
